// FR-ADMIN/src/app/agent/model/agent.model.ts
import { Agence } from "src/app/agence/model/agence.model";

export class Agent {
	id: number | any;
	nom: string | any;
	prenom: string | any;
	cin: string | any;
	adresse: string | any;
	telephone: string | any;
	email: string | any;
	agence: Agence | any;
	username: string | any;
	password: string | any;
}